Final state.

+++ Assets/Scripts/setupScene.cs
@@ -109,114 +109,32 @@ public class setupScene : MonoBehaviour{
     public void noCalibration(){
         Debug.Log("[SETUP SCENE] No Calibration has been selected. Loading saved information.");
         string[] planeCalibDatText = System.IO.File.ReadAllLines(Application.dataPath + "/Resources/planeCalibData.txt");        
-        if (planeCalibDatText.Length != 12){
+        if (planeCalibDatText.Length != 6){
             Debug.LogError("[SETUP SCENE] 'No calibration' has been selected, but no valid text file has been read.");
             SceneManager.UnloadSceneAsync(SceneManager.GetSceneByName("CalibrateOrNot"));
             SceneManager.LoadScene("SelectCalibrationTarget", LoadSceneMode.Additive);
         }else{
             Vector3 lowerLeft = new Vector3();
             Vector3 upperRight = new Vector3();
-            Vector3 additionalLowerRight = new Vector3();
-            Vector3 heightDeviations = new Vector3();
             lowerLeft.x = float.Parse(planeCalibDatText[0]);
             lowerLeft.y = float.Parse(planeCalibDatText[1]);
             lowerLeft.z = float.Parse(planeCalibDatText[2]);
             upperRight.x = float.Parse(planeCalibDatText[3]);
             upperRight.y = float.Parse(planeCalibDatText[4]);
             upperRight.z = float.Parse(planeCalibDatText[5]);
-            additionalLowerRight.x = float.Parse(planeCalibDatText[6]);
-            additionalLowerRight.y = float.Parse(planeCalibDatText[7]);
-            additionalLowerRight.z = float.Parse(planeCalibDatText[8]);
-            heightDeviations.x = float.Parse(planeCalibDatText[9]);
-            heightDeviations.y = float.Parse(planeCalibDatText[10]);
-            heightDeviations.z = float.Parse(planeCalibDatText[11]);
 
             Debug.Log("[LOADING CALIBRATION DATA] lowerLeft: " + lowerLeft);
-            Debug.Log("[LOADING CALIBRATION DATA] upperRight: " + upperRight);
-            Debug.Log("[LOADING CALIBRATION DATA] additionalLowerRight: " + additionalLowerRight);
-            Debug.Log("[LOADING CALIBRATION DATA] heightDeviations: " + heightDeviations);
+            Debug.Log("[LOADING CALIBRATION DATA] upperRight: " + upperRight);            
 
-            calibrationDone(lowerLeft, upperRight, additionalLowerRight, heightDeviations);
+            calibrationDone(lowerLeft, upperRight);
 
             SceneManager.UnloadSceneAsync(SceneManager.GetSceneByName("CalibrateOrNot"));
             SceneManager.LoadScene("SetScale", LoadSceneMode.Additive);
         }
     }
 
-    //// Is called when table calibration finishes (in TableCalibration.cs)
-    //public void calibrationDone(Vector3 lowerLeft, Vector3 upperRight){
-    //    // Make marker positions available globally
-    //    calibratedLL = lowerLeft;
-    //    calibratedUR = upperRight;
-
-    //    GameObject alreadyCalibrated = GameObject.Find("TablePlane");
-    //    if (alreadyCalibrated != null)
-    //        Destroy(alreadyCalibrated);
-
-    //    // Create plane (table surface)
-    //    table = GameObject.CreatePrimitive(PrimitiveType.Plane);
-    //    table.name = "TablePlane";
-    //    table.transform.parent = parent.transform;
-    //    float width = Math.Abs(calibratedUR.x - calibratedLL.x);
-    //    float height = Math.Abs(calibratedUR.z - calibratedLL.z);
-    //    Vector3 position = new Vector3(
-    //                                    (calibratedLL.x + calibratedUR.x) / 2,
-    //                                    ((calibratedLL.y + calibratedUR.y) / 2) + planeHeightOffset,
-    //                                    (calibratedLL.z + calibratedUR.z) / 2
-    //                                  );
-    //    table.transform.position = position;
-    //    table.transform.localScale = new Vector3(width / 10, 1, height / 10);
-
-    //    GameObject tableMenuParent = GameObject.Find("TableMenuParent");
-    //    tableMenuParent.transform.parent    = table.transform;
-    //    tableMenuParent.transform.position = new Vector3(position.x  +width / 2 , position.y, position.z - height / 2  );
-
-    //    calibDone = true;        
-    //}
-
-    //private float calculatePlaneRotation(Vector3 lowerLeft, Vector3 upperRight, Vector3 additionalLowerRight, Vector3 heightDeviations){
-    //    heightDeviations = new Vector3(1, 1, 1);
-    //    workspacePlane = new Plane();
-    //    Vector3 planeUR = new Vector3(upperRight.x, upperRight.y * heightDeviations.y, upperRight.z);
-       
-    //    calibratedUR = planeUR;        
-    //    Vector3 planeAdditionalLR = new Vector3(additionalLowerRight.x, additionalLowerRight.y * heightDeviations.z, additionalLowerRight.z);
-    //    Vector3 planeLL = new Vector3(lowerLeft.x, lowerLeft.y * heightDeviations.x, lowerLeft.z);
-        
-    //    calibratedLL = planeLL;
-    //    workspacePlane.Set3Points(planeUR, planeAdditionalLR, planeLL);
-    //    Vector3 planeNormalVector = workspacePlane.normal;
-
-    //    Debug.DrawLine(planeLL, planeAdditionalLR, Color.green, 3600f, false);
-    //    Debug.DrawLine(planeAdditionalLR, planeUR, Color.green, 3600f, false);
-    //    Debug.DrawLine(planeUR, planeLL, Color.green, 3600f, false);
-
-    //    //Plane planeZ0 = new Plane();
-    //    //Vector3 planeZ0UR = new Vector3(upperRight.x, additionalLowerRight.y, upperRight.z);
-    //    //Vector3 planeZ0AdditionalLR = new Vector3(additionalLowerRight.x, additionalLowerRight.y, additionalLowerRight.z);
-    //    //Vector3 planeZ0LL = new Vector3(lowerLeft.x, additionalLowerRight.y, lowerLeft.z);
-    //    //planeZ0.Set3Points(planeZ0UR, planeZ0AdditionalLR, planeZ0LL);
-    //    //Vector3 planeZ0NormalVector = planeZ0.normal;
-
-    //    Plane planeZ0 = new Plane();
-    //    //Vector3 planeZ0UR = new Vector3(upperRight.x, additionalLowerRight.y, upperRight.z);
-    //    Vector3 planeZ0AdditionalLR = new Vector3(upperRight.x, (upperRight.y + lowerLeft.y) / 2, lowerLeft.z);
-    //    //Vector3 planeZ0LL = new Vector3(lowerLeft.x, additionalLowerRight.y, lowerLeft.z);
-    //    planeZ0.Set3Points(planeUR, planeZ0AdditionalLR, planeLL);
-    //    Vector3 planeZ0NormalVector = planeZ0.normal;
-
-    //    Debug.DrawLine(planeLL, planeZ0AdditionalLR, Color.red, 3600f, false);
-    //    Debug.DrawLine(planeZ0AdditionalLR, planeUR, Color.red, 3600f, false);
-    //    Debug.DrawLine(planeUR, planeLL, Color.red, 3600f, false);
-
-    //    float dotProduct = Vector3.Dot(planeNormalVector, planeZ0NormalVector);
-    //    float lengthProduct = planeNormalVector.magnitude * planeZ0NormalVector.magnitude;
-    //    float angleInRad = (float)Math.Acos(dotProduct / lengthProduct);
-    //    return (float)(angleInRad * (180.0 / Math.PI));
-    //}
-
     // Is called when table calibration finishes (in TableCalibration.cs)
-    public void calibrationDone(Vector3 lowerLeft, Vector3 upperRight, Vector3 additionalLowerRight, Vector3 heightDeviations){
+    public void calibrationDone(Vector3 lowerLeft, Vector3 upperRight){
         calibratedLL = lowerLeft;
         calibratedUR = upperRight;
 
@@ -256,8 +174,10 @@ public class setupScene : MonoBehaviour{
 
     private void deleteSavesDir(){
         string path = Application.dataPath + "/Resources/saves/";
-        Directory.Delete(path, true);        
-        Directory.CreateDirectory(path);
+        if (Directory.Exists(path)) { 
+            Directory.Delete(path, true);
+            Directory.CreateDirectory(path);
+        }
     }
 
     void Start() {
